将base64升级为VS2008

+++ trunk/Base64编解码器/Base64/Base64Dlg.cpp
@@ -404,7 +404,7 @@ void CBase64Dlg::OnBnClickedDecodetofile()
 		}
 		CString pathName = fileDlg.GetPathName();
 		CFile file(pathName,CFile::modeCreate | CFile::modeWrite);
-		file.Write(sourceByte,sourceLen);
+		file.Write(sourceByte,(UINT)sourceLen);
 		file.Close();
 		delete[] sourceByte;
 	}
